+++ examples/voice.rs
@@ -16,7 +16,7 @@ fn main() {
     let mut synth = Synth::new();
     // Generate audio samples.
     synth.gen(audio.sink(..), |fc| {
-        fc.freq(440.0).abs().gain(fc.freq(440.0).sine().into())
+        fc.freq(440.0).abs().gain(fc.freq(440.0).sine())
     });
 
     // Write chord to file